Use .Site.Author.name in footer instead of .Site.Params.Author
Hugo provides an Author site variable that can be set in the site
configuration. Instead of setting the author variable in the params,
section from this commit on, you have to use the name variable in the
Author section. This adheres more to Hugo standards. See README.MD for
an example.